Add stop-on-signals and auto-cancellation
When workflows are cancelled, any started builds are left running. This
is not usually desired, for example if using workflow concurrency[^1] to
cancel redundant builds. In cases like this, we'd want to also stop any
triggered builds.

This commit adds a `stop-on-signals` input with a default value of
`SIGINT` and updates the action to listen for the listed signal(s) and
stop any already-started build in response. This accomplishes the
desired behavior in the common case.

[^1]: https://docs.github.com/en/actions/using-jobs/using-concurrency#example-using-concurrency-to-cancel-any-in-progress-job-or-run