Updated githubInfo function to work for non github.com urls
The original code has github.com hardcoded. This does not work for github enterprise servers where the hostname is different.

Also had to fix the owner/repo extraction to account for extra slashes in a remote host with weird pathing, and to check for the ":" which is required in an SSH url.